Ready to merge

+++ src/breakpoint.h
@@ -5,9 +5,10 @@
 #include <sys/ptrace.h>
 #include <stdint.h>
 #include <stdlib.h>
+#include <stdio.h>
 
 
 
-uint32_t add_breakpoint(pid_t pid, uint32_t addr);
-void remove_breakpoint(pid_t pid, uint32_t addr, uint32_t old_code);
+uint64_t add_breakpoint(pid_t pid, uint64_t addr);
+void remove_breakpoint(pid_t pid, uint64_t addr, uint64_t old_code);
